+++ HISTORY.rst
@@ -1,3 +1,10 @@
 =========
 Changelog
 =========
+
+0.0.1
+-----
+
+Initial release. Use Visma API objects in an ORM:y way. get, create, update
+and delete function. All objects added after editing output from
+swagger-marshmallow-codegen. But only some is working properly.